third_party/WebKit/LayoutTests/fast/multicol/vertical-rl/offset-top-and-left-nested.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div style="position:relative; writing-mode:vertical-rl; columns:3; column-fill:auto; column-gap:10px; width:100px; height:350px;">
    <div style="columns:2; column-fill:auto; column-gap:10px; width:200px;">
        <div style="width:250px;"></div>
        <div id="in-second-row" style="width:20px;"></div>
    </div>
</div>

<script>
    test(function() {
        assert_equals(document.getElementById('in-second-row').offsetLeft, 30);
        assert_equals(document.getElementById('in-second-row').offsetTop, 120);
    }, "offsetLeft and offsetTop in vertical-rl in nested multicol");
</script>
